authentication/auth_requirements.txt
fastapi>=0.68.0,<0.100.0
uvicorn>=0.15.0,<0.25.0
pydantic>=1.8.2,<2.0.0
PyJWT>=2.1.0,<3.0.0
pymongo>=4.0.0,<5.0.0
python-dotenv>=0.19.0,<1.0.0